+++ src/service-worker.js
@@ -1,2 +1,1 @@
 export default () => 'future of libraries is open';
-